make `firebase` the last command in the script
The exit code of the script should be the exit code of the `firebase ` command. The easiest way to achieve that is to have `firebase` be the last command in the script.

I've removed the (conditional) token deprecation warning output at the end of the script, because it's a duplicate, the warning is still given before the `firebase` command is executed.